requirements.txt
django==4.0
django-filter
django-cors-headers==3.10.1
djangorestframework==3.13.1
django-cors-middleware==1.5.0
requests==2.26.0
pylint==2.12.2
markdown==3.3.6
python-dotenv==0.19.2